🔧 common 모듈 추가 (#6)
* 🔥 앱 모듈 내 di 제거

* 🔧 Common module 내 di 추가
