+++ common/build.gradle.kts
@@ -35,5 +35,8 @@ android {
 }
 
 dependencies {
+    implementation(project(":domain"))
+    implementation(project(":data"))
+    
     applyHilt()
 }
